% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_data.R
\docType{data}
\name{agent_id}
\alias{agent_id}
\title{Agent Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.
}
\usage{
agent_id
}
\description{
Data for assigning agents to their UUID's
}
\details{
\describe{
\item{agent}{Agent Name}
\item{uuid}{ID of agent}

}
}
\keyword{datasets}
