% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunFunctions.R
\name{allGunStats}
\alias{allGunStats}
\title{Creates a visualization about all guns in respect to a certain statistic}
\usage{
allGunStats(statistic)
}
\arguments{
\item{gun}{a statistic in a string (i.e. "fireRate", "magazineSize", "equipTimeSeconds", "reloadTimeSeconds")}
}
\value{
A boxplot of all of the values for each gun
}
\description{
Creates a visualization about all guns in respect to a certain statistic
}
