% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentFunctions.R
\name{getAgentAbilities}
\alias{getAgentAbilities}
\title{Gets description of inputted agent's abilities}
\usage{
getAgentAbilities(agent)
}
\arguments{
\item{agent}{name of agent in a string}
}
\value{
A data frame of agent info
}
\description{
Gets description of inputted agent's abilities
}
