% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agentFunctions.R
\name{getAgentInfo}
\alias{getAgentInfo}
\title{Gets description of inputted agent}
\usage{
getAgentInfo(agent)
}
\arguments{
\item{agent}{name of agent (capitalized)}
}
\value{
A string that is an agent's description
}
\description{
Gets description of inputted agent
}
