% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunFunctions.R
\name{getGunDamage}
\alias{getGunDamage}
\title{Gets damage stats about inputted gun}
\usage{
getGunDamage(gun)
}
\arguments{
\item{gun}{name of gun in a string}
}
\value{
A data frame of damage information about the gun
}
\description{
Gets damage stats about inputted gun
}
