% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunFunctions.R
\name{getGunStats}
\alias{getGunStats}
\title{Gets basic stats about inputted gun}
\usage{
getGunStats(gun)
}
\arguments{
\item{gun}{name of gun in a string}
}
\value{
A data frame of information about the gun
}
\description{
Gets basic stats about inputted gun
}
