% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFunctions.R
\name{getMapInfo}
\alias{getMapInfo}
\title{Gets map information about inputted map}
\usage{
getMapInfo(map)
}
\arguments{
\item{map}{name of map in a string}
}
\value{
A dataframe that lists all of the locations on the map
}
\description{
Gets map information about inputted map
}
