% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gun_data.R
\docType{data}
\name{gun_id}
\alias{gun_id}
\title{Gun Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 17 rows and 2 columns.
}
\usage{
gun_id
}
\description{
Data for assigning VALORANT guns to their UUID's
}
\details{
\describe{
\item{gun}{Gun Name}
\item{uuid}{ID of gun}

}
}
\keyword{datasets}
