% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunFunctions.R
\name{gunFireLevel}
\alias{gunFireLevel}
\title{Creates a visualization about all guns in respect to a certain statistic}
\usage{
gunFireLevel(wall_pen_level)
}
\arguments{
\item{wall_pen_level}{("Low", "Medium", "High")}
}
\value{
box plot of wall penetration levels and their corresponding fire rates of all guns
}
\description{
Creates a visualization about all guns in respect to a certain statistic
}
