% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\docType{data}
\name{map_id}
\alias{map_id}
\title{Map Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\usage{
map_id
}
\description{
Data for assigning maps to their UUID's
}
\details{
\describe{
\item{map}{Map Name}
\item{uuid}{ID of map}

}
}
\keyword{datasets}
